Moved all logic to model, separated CRUD from reest of the logic.
