+++ src/main/java/citygenerator/model/DataLayer/BussinesLogic/markov/Generator.java
@@ -1,4 +1,4 @@
-package citygenerator.markov;
+package citygenerator.model.DataLayer.BussinesLogic.markov;
 
 import citygenerator.util.ListTools;
