+++ src/main/java/citygenerator/model/DataLayer/BussinesLogic/markov/MarkovChain.java
@@ -1,4 +1,4 @@
-package citygenerator.markov;
+package citygenerator.model.DataLayer.BussinesLogic.markov;
 
 import java.util.ArrayList;
 import java.util.Stack;
