+++ src/main/java/citygenerator/model/DataLayer/BussinesLogic/namegenerator/NameGenerator.java
@@ -1,0 +1,11 @@
+package citygenerator.model.DataLayer.BussinesLogic.namegenerator;
+
+import java.util.List;
+
+public interface NameGenerator {
+
+    void setSeed(Long seed);
+    Long getSeed();
+    String getName();
+    List<String> getNames(Integer count);
+}
