--- src/main/java/citygenerator/model/DataLayer/namegenerator/NameGenerator.java
@@ -1,8 +1,0 @@
-package citygenerator.model.DataLayer.namegenerator;
-
-public interface NameGenerator {
-
-    void setSeed(String seed);
-    String getSeed();
-    String getName();
-}
